Refactor rte_general functions to static methods
